pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["OlegKhiretdinov <oleg@oleg.win>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
python-dotenv = "^1.0.0"
dj-database-url = "^2.1.0"
django-bootstrap5 = "^23.3"
gunicorn = "^21.2.0"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
psycopg2-binary = "^2.9.9"
django-filter = "^23.5"
rollbar = "^0.16.3"
django-extensions = "^3.2.3"
djangorestframework = "^3.14.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
ipython = "^8.20.0"
coverage = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
